feat: Lotto 구현
